Fix namespacing. Fix constraints on JsHashMap. Relocated FindOptions to room.
